requirements.txt
# 高熵合金论文抽取 Pipeline 依赖
# Python >= 3.10

langextract>=1.0.0
pydantic>=2.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
